Синхронизация с rusAcis 3.1 part 2.

+++ java/ru/privetdruk/l2jspace/gameserver/model/item/DropData.java
@@ -49,7 +49,7 @@ public class DropData {
     /**
      * @return the chance to have a drop, under a 1.000.000 chance.
      */
-    public double getChance() {
+    public int getChance() {
         return _chance;
     }
 }
