+++ java/ru/privetdruk/l2jspace/gameserver/scripting/quest/Q327_RecoverTheFarmland.java
@@ -1,16 +1,15 @@
 package ru.privetdruk.l2jspace.gameserver.scripting.quest;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import ru.privetdruk.l2jspace.common.random.Rnd;
-
 import ru.privetdruk.l2jspace.gameserver.enums.QuestStatus;
 import ru.privetdruk.l2jspace.gameserver.model.actor.Creature;
 import ru.privetdruk.l2jspace.gameserver.model.actor.Npc;
 import ru.privetdruk.l2jspace.gameserver.model.actor.Player;
 import ru.privetdruk.l2jspace.gameserver.scripting.Quest;
 import ru.privetdruk.l2jspace.gameserver.scripting.QuestState;
+
+import java.util.HashMap;
+import java.util.Map;
 
 public class Q327_RecoverTheFarmland extends Quest {
     private static final String QUEST_NAME = "Q327_RecoverTheFarmland";
@@ -220,7 +219,7 @@ public class Q327_RecoverTheFarmland extends Quest {
                 int n = player.getInventory().getItemCount(i);
                 if (n > 0) {
                     takeItems(player, i, n);
-                    rewardExpAndSp(player, (long) n * EXP_REWARD.get(i), 0);
+                    rewardExpAndSp(player, n * EXP_REWARD.get(i), 0);
                     isRewarded = true;
                 }
             }
